Use ReportbackTransformer instead of PhoenixGalleryTransformer
